Generate warning comments
