+++ src/reportSizeImpactIntoGithubPullRequest.js
@@ -11,7 +11,6 @@ import { getPullRequestCommentMatching } from "./internal/getPullRequestCommentM
 import { createPullRequestComment } from "./internal/createPullRequestComment.js"
 import { updatePullRequestComment } from "./internal/updatePullRequestComment.js"
 import { generatePullRequestCommentString } from "./internal/generatePullRequestCommentString.js"
-import { compareTwoSnapshots } from "./internal/compareTwoSnapshots.js"
 import { jsenvFormatSize } from "./internal/jsenvFormatSize.js"
 
 const regexForMergingSizeImpact = /<!-- Generated by @jsenv\/github-pull-request-filesize-impact -->/
@@ -105,44 +104,11 @@ ${headSnapshotFileContent}
     const baseVersionnedSnapshot = JSON.parse(baseSnapshotFileContent)
     const headVersionnedSnapshot = JSON.parse(headSnapshotFileContent)
 
-    let baseSnapshot
-    const headSnapshot = headVersionnedSnapshot.snapshot
-    const warnings = []
-
-    // this is an indirect way of detecting that there is an error while generating snapshot
-    // There is || echo "{}" > ../snapshot.base.json in size-impact.yml
-    if (Object.keys(baseVersionnedSnapshot).length === 0) {
-      baseSnapshot = {}
-      warnings.push(
-        `**Warning:** Only \`${pullRequestHead}\` files are taken into account below.
-It happens because \`${pullRequestBase}\` files are missing.
-This occurs when there is an error executing \`@jsenv/github-pull-request-filesize-impact\` scripts.
-It's likely because scripts are not in \`${pullRequestBase}\` branch.
-This is normal when adding \`@jsenv/github-pull-request-filesize-impact\` for the first time.`,
-      )
-    } else {
-      const baseVersion = baseVersionnedSnapshot.version
-      const headVersion = headVersionnedSnapshot.version
-
-      if (baseVersion === headVersion) {
-        baseSnapshot = baseVersionnedSnapshot.snapshot
-      } else {
-        baseSnapshot = {}
-        logger.warn(`base snapshot ignored because different version.`)
-        warnings.push(
-          `**Warning:** Only \`${pullRequestHead}\` files are taken into account below.
-It happens because versions of \`@jsenv/github-pull-request-filesize-impact\` are too different on \`${pullRequestBase}\` and \`${pullRequestHead}\`.`,
-        )
-      }
-    }
-
-    const snapshotComparison = compareTwoSnapshots(baseSnapshot, headSnapshot)
-
     const pullRequestCommentString = generatePullRequestCommentString({
       pullRequestBase,
       pullRequestHead,
-      snapshotComparison,
-      warnings,
+      baseVersionnedSnapshot,
+      headVersionnedSnapshot,
       formatSize,
       commentSections,
       generatedByLink,
